+++ CHANGELOG.md
@@ -13,6 +13,8 @@ and this project adheres to
 - Pointers may now be used in if conditions, tenary conditions and as operands in logical AND and OR expressions
   - [#3656](https://github.com/bpftrace/bpftrace/pull/3656)
 #### Changed
+- `probe` builtin is now represented as a string type
+  - [#3638](https://github.com/bpftrace/bpftrace/pull/3638)
 #### Deprecated
 #### Removed
 #### Fixed